Add enum_from_primitive to CLStatus
An alternative would be to change the C bindings to return CLStatus, but
we probably want to keep the binding signatures as close as possible to
the original C and handle the translation to Rust Land ourselves. Also,
if for whatever reason openCL returns an invalid cl_int status, we
probably shouldn't start triggering undefined behaviour.